cleanup and made npm install + prepare build more durable, can be called from anywhere in repo
